requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
fastapi==0.110.2
greenlet==3.0.3
h11==0.14.0
idna==3.7
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
